Change location of BUILDTOOLS_DIR
Make the buildtools Git submodule-friendly by changing the BUILTOOLS_DIR
variable: projects which use the base config file should now expect to
find the buildtools in a subdirectory of the parent project's source
root.

This should make individual projects more robust in the face of changes
to buildtools, because it means that they can pin the version of
buildtools that they use to a specific, "known good" commit.

The developer of a project using buildtools can chose to integrate
upstream changes when he/she wishes, rather than being forced to
integrate and fix any breakage just because upstream happened to move in
the intervening interval.

It should also make it easier to make initial clones of projects using
buildtools because now users can just pull down the parent project
rather than having to worry about performing multiple clones and
ensuring that each repository is in the appropriate relative location.

This is actually quite a large change so at the time of writing I've
split the repository into two branches: this commit is appearing first
on the "master" branch, while the "maint" branch is being left as it was
before.

Signed-off-by: Wincent Colaiuta <win@wincent.com>