ReleaseNotes.sh: handle submodules ahead of annotated tags
These changes only affect behaviour when the "--tag-prefix=" switch is
passed in, which is typically when working within submodules.

Things worked correctly when the submodule HEAD was at an annotated tag,
but broke down when it was somewhere ahead of one.

Without the "--tag-prefix=" switch we correctly produced results like:

  "Changes from 2.0 to 2.0-13-gf96082e"

But with the switch we instead produced results like:

  "Changes from 2.0 to 2.0"

And we didn't list any changes at all.

Now we list changes since the last annotated tag in that case, bring the
behaviour into line for both cases.

Note that this commit does add one "BUG" comment, because I noticed
while testing that this will break down in repos with no annotated tags
("git describe" bails in such repos). There are probably other issues
too when running the script in repos whose HEAD points to older TAGS (I
suspect it will always just produce release notes for the latest
version), so there is still room for some improvement.

Signed-off-by: Wincent Colaiuta <win@wincent.com>