ReleaseNotes.sh: handle first release
Our "git describe HEAD^" call fails for the first tagged release because
there are no prior tags.

Catch the error and just use the first root commit instead.

This is not the only place where we use "git describe", but it is the
most likely place where problems will crop up in practice (because
usually there is at least _one_ annotated tag in the repo if we are
running this script).

Signed-off-by: Wincent Colaiuta <win@wincent.com>