Add git-last-tag.sh script
This script can be used to help produce incremental release notes. It
reports the last-found tag prior to the current HEAD commit. That is,
if the current HEAD is tagged as 1.1, and the previous tagged release
was 1.0, it will print 1.0.

If some commits have been made since the last release and the current
HEAD does not correspond to a tag then it will print the last found
tag.

Signed-off-by: Wincent Colaiuta <win@wincent.com>