Replace ReleaseNotes.sh script with rewrite, git-release-notes.txt
The old ReleaseNotes.sh script was intended to be run once for the main
project, and then again for each submodule within the project.

The new git-release-notes.sh script need only be run on the main
project, and it will automatically incorporate changelogs for any
top-level submodules. As I don't have any submodules within submodules
at this stage, nor any submodules in non-top-level subdirectories, this
is not a recursive operation.

Signed-off-by: Wincent Colaiuta <win@wincent.com>