Lay separate build-intermediates folder to rest
I'm currently trying to simplify the build folder layout to handle some
new submodule workflows, so this is one way to cut down the number of
moving parts. The OBJROOT folder is actually fairly neat and tidy, so
merging its contents into SYMROOT is not going to clutter things up.

Signed-off-by: Wincent Colaiuta <win@wincent.com>