Set LD_DYLIB_INSTALL_NAME using @executable_path
This should allow applications to locate the framework when they embed
it. In order to install such a framework to /Library/Frameworks/, the
InstallFramework.sh script can be used.

I haven't used it for some time, but I believe the
installed-framework-target.xcconfig file is probably in a broken state
right now; without the INSTALL_PATH and DSTROOT stuff it should be
possible to build a framework suitable for installation using the
default Xcode build setting for LD_DYLIB_INSTALL_NAME.

Signed-off-by: Wincent Colaiuta <win@wincent.com>