Drop GCC_SYMBOLS_PRIVATE_EXTERN setting
This always used to work for me with my 32-bit builds, but now that I am
targeting 64-bit it's biting me. Basically, it hides Objective-C classes
now, which means you can't link to frameworks built with this build
setting active.

A couple resources I turned up in my search for info on this:

  http://lists.apple.com/archives/xcode-users/2009/Dec/msg00178.html
  (post from a guy also surprised by the behavior change in the move
  from 32-bit to 64-bit)

  http://www.cocoabuilder.com/archive/cocoa/217622-how-to-export-class-while-the-framework-gcc-symbols-private-extern-is-yes.html
  (thread which suggests that this might actually be a Clang/LLVM thing
  only, as GCC apparently thinks that visibility is "not applicable to
  Objective-C classes")

Whatever the case, I'm losing this build setting. It was only there to
make reverse-engineering a tiny bit harder, and I'm much more interested
in openness and things just working, than obfuscation and secrecy
nowadays.

Signed-off-by: Wincent Colaiuta <win@wincent.com>